+++ rbx_binary/src/deserializer/state.rs
@@ -295,13 +295,24 @@ impl<'db, R: Read> DeserializerState<'db, R> {
         let mut referents = vec![0; number_instances as usize];
         chunk.read_referent_array(&mut referents)?;
 
+        let prop_capacity = self
+            .deserializer
+            .database
+            .classes
+            .get(type_name.as_str())
+            .map(|class| class.default_properties.len())
+            .unwrap_or(0);
+
         // TODO: Check object_format and check for service markers if it's 1?
 
         for &referent in &referents {
             self.instances_by_ref.insert(
                 referent,
                 Instance {
-                    builder: InstanceBuilder::new(&type_name),
+                    builder: InstanceBuilder::with_property_capacity(
+                        type_name.as_str(),
+                        prop_capacity,
+                    ),
                     children: Vec::new(),
                 },
             );